+++ src/data-flow/service-registration.js
@@ -2,6 +2,7 @@ import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { CommandFlowDescriptor } from './commands/command-flow-descriptor.js';
 import { EventFlowDescriptor } from './events/event-flow-descriptor.js';
 import { ExecutionOverviewDescriptor } from './execution/execution-overview-descriptor.js';
+import { ServiceIntegrationDescriptor } from './integrations/service-integration-descriptor.js';
 import { QueryFlowDescriptor } from './queries/query-flow-descriptor.js';
 import { WorkflowOrchestrationDescriptor } from './workflows/workflow-orchestration-descriptor.js';
 
@@ -14,6 +15,7 @@ export function addDataFlow(services) {
   services.registerSingleton('EventFlowDescriptor', () => new EventFlowDescriptor());
   services.registerSingleton('ExecutionOverviewDescriptor', () => new ExecutionOverviewDescriptor());
   services.registerSingleton('QueryFlowDescriptor', () => new QueryFlowDescriptor());
+  services.registerSingleton('ServiceIntegrationDescriptor', () => new ServiceIntegrationDescriptor());
   services.registerSingleton('WorkflowOrchestrationDescriptor', () => new WorkflowOrchestrationDescriptor());
 
   return services;